extract config into config.rs
